+++ changelog.md
@@ -3,6 +3,8 @@
 - Images are now offline and optimized in webp format
 - Fixes for missing data for previous years
 - Optimization of data format
+- Links for art added
+- Links open in browser
 
 ### 1.12
 - Added "now" to events to show events starting in the next hour (or running for up to 20 mins)